feat: 薪资模板改造 - 模板CRUD/员工绑定 + 按模板算工资/加班费 + 人效按模板 + 考勤半天标记
- 新增薪资模板表与CRUD，员工必选模板(隐藏日薪/底薪)
- 算工资按模板类型(hourly/daily/monthly)，支持半天出勤/缺勤/请假
- 加班费按员工模板半小时费算(分钟向上取整半小时)，无模板拦截
- 人效人工成本改从薪资模板取，无模板按0跳过
- 考勤日历支持半天/半天旷工(红色)标记

+++ backend/app/models/payroll.py
@@ -19,10 +19,10 @@ class PayrollRecord(Base):
 
     # Attendance breakdown
     total_days_in_month = Column(Integer, default=0)
-    attendance_days = Column(Integer, default=0)
-    leave_days = Column(Integer, default=0)
-    rest_days = Column(Integer, default=0)
-    absence_days = Column(Integer, default=0)
+    attendance_days = Column(Float, default=0)
+    leave_days = Column(Float, default=0)
+    rest_days = Column(Float, default=0)
+    absence_days = Column(Float, default=0)
 
     # Salary components
     employee_status = Column(String(20), default="trial")  # trial / regular
